requirements.txt
playwright
pytest
pyyaml
python-dotenv
